Change Google Analytics to also include an indicator as to whether results were returned
The Google Analytics code has been moved to the results page so that we can detect if results were
found. It then adds another parameter to the virtual page in Google Analytics.
There are a few 'limitations' with this technique.
Firstly, any searches for existing title numbers will not go to the result page and therefore will
not record these searches.
Secondly, the browser back button can cause successful search terms to record no results found if
the previous result had no results.